# server/agent-template/public/openapi.yaml
openapi: 3.1.0
info:
  title: eCG Agent Dashboard API
  version: "1.0.0"
  description: |
    This is the API your generated dashboard talks to. Every call goes through
    `ecgApi` (see `src/lib/ecgClient.ts`) — never raw `fetch` — and every route
    below is bridged server-side to a real eCG Agent MCP tool call. You never
    need an MCP client library, a session handshake, or the raw MCP tool names:
    this REST surface *is* the MCP surface, translated into ordinary HTTP.

    ## Two connection modes
    Every dashboard is connected one of two ways, and that decides which
    endpoints actually work:

    - **MCP key** (the current, default way to connect — paste an `ecg_...`
      key from `app.ecomgear.ai/org/api-keys`): every endpoint tagged **MCP**
      below works. Endpoints tagged **Portal-only** return `501 Not Implemented`
      — there is no MCP tool behind them.
    - **Portal account** (an older, retired onboarding path some existing
      dashboards still use): both **MCP** and **Portal-only** endpoints work,
      because the connection carries a full portal session, not just an
      agent-management key.

    If you're building a custom feature, assume **MCP-only** unless you know
    otherwise — it's the common case, and every built-in page in this
    dashboard is written to degrade gracefully when a Portal-only call 501s
    (see `isUnsupported()` in `src/lib/ecgClient.ts`).

    ## Authentication
...

...

...

...
  contact:
    name: eCG Agent dashboard template
...
servers:
...
security:
  - dashboardAccess: []
  - ownerSession: []
...
components:
  securitySchemes:
    dashboardAccess:
      type: apiKey
      in: header
      name: x-dashboard-access
      description: Issued by `POST /api/v1/ecg-access`; used automatically by `ecgClient.ts` for anonymous dashboard visitors.
    ownerSession:
      type: http
      scheme: bearer
      description: Supabase session JWT for the eComGear account that owns this project; attached automatically when you're logged in.
...
paths:
...
  /connectors/org/{id}/test:
    parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
    post:
      tags: [Connectors]
      summary: Verify a connector actually works
      x-mcp-support: MCP
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { type: object, properties: { note: { type: string } } }
...
  /knowledge/upload:
    post:
      tags: [Knowledge]
      summary: Upload a knowledge document
      x-mcp-support: MCP
      requestBody:
        content:
          multipart/form-data:
            schema: { type: object, properties: { file: { type: string, format: binary } } }
      responses:
        '200': { description: Uploaded }
...
  /ai-chat:
    post:
      tags: [Assistant]
      summary: Plain LLM passthrough (NOT what the built-in Assistant uses)
      description: >-
        A tool-less, direct call to whichever LLM provider this org has
        configured. The dashboard's own "Assistant" tab does NOT call this —
        it calls `POST /api/v1/ecg-chat` (a different, sibling endpoint, not
        under this proxy) via the top-level `chat()` function in
        `ecgClient.ts`, which can actually invoke tools (list/create/run
        agents, approve posts, etc.). Reach for this endpoint only if you
        want a raw completion with zero tool access; extend `ChatPage.tsx` /
        `ecg-chat.routes.ts` instead of duplicating agentic chat logic here.
      x-mcp-support: MCP
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [messages]
              properties:
                messages:
                  type: array
                  items:
                    type: object
                    properties: { role: { type: string, enum: [user, assistant, system] }, content: { type: string } }
      responses:
        '200': { description: OK }